Move the background into the row.
